Disable the ability to downgrade the generator during a meltdown
Currently, it's possible to downgrade the generator during a meltdown, preventing the meltdown entirely. I'm assuming this was not intended as this feels like a "cheesy" to avoid it.

The downgrade dialog checks if the generator is during a meltdown, and if it is - display a message that it's disallowed. If the generator is not in a meltdown the downgrade will happen normally.

I've considered adding this check to downgrade gizmo instead, however I've realized that it would still be possible to downgrade a generator having a meltdown if the dialog was opened before a meltdown happened.