Upgrade gizmo checks for frame before placing blueprint
The current gizmo only checks if a blueprint is present at the same location, but not the building frame. I've changed it so both are checked. On top of that, the upgrade from chemfuel fortified to uranium powered did not have a check for the blueprint, likely due to small oversight.

To avoid copy-pasting the same snippet of the code everywhere, I've made a new method that handles placing the blueprint (`Utils.PlaceDistinctBlueprint`). This should hopefully prove useful if some changes or improvements are ever needed to this code since all the upgrade gizmo actions will be handled in a single place now.

As a slight side note - I could optimize the method placing the blueprint to only use 1 loop rather than 2 (the `ThingAt` call). However, given that this is a method that will rarely be called, and only from gizmos - I've decided that there's likely no point in doing it.